+++ Projects/babyMoa/babyMoa/App/AppDelegate.swift
@@ -8,6 +8,7 @@ import UIKit
 
 class AppDelegate: NSObject, UIApplicationDelegate {
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey : Any]? = nil) -> Bool {
+        print(UserToken.accessToken)
 //        // MARK: UserDefault 데이터 삭제 필요 시 주석 해제하세요. refresh token 로직 추가하면 이렇게 할 필요 없습니다.
 //        if let bundleID = Bundle.main.bundleIdentifier {
 //                UserDefaults.standard.removePersistentDomain(forName: bundleID)
